feat(60-01): create market-schemas.ts and clean up schemas.ts
- Add market-schemas.ts with ensureMarketTables() — 4 tables: daily, context, intraday, _sync_metadata
- All columns created upfront with correct PKs per plan spec
- Remove ensureMarketDataTables() from schemas.ts (old spx_daily/spx_15min/vix_intraday tables)
- Remove market._sync_metadata from ensureSyncTables() (now lives in market.duckdb)
- Export tableExists() from schemas.ts for use in later phases
- Update db/index.ts: remove ensureMarketDataTables export, add ensureMarketTables + tableExists exports
