+++ packages/mcp-server/src/db/index.ts
@@ -1,7 +1,8 @@
 /**
  * DuckDB module exports
  *
- * Provides connection management and schema definitions for the analytics database.
+ * Provides connection management and schema definitions for the analytics database
+ * (analytics.duckdb) and the market database (market.duckdb).
  */
 
 export { getConnection, closeConnection, isConnected, upgradeToReadWrite, downgradeToReadOnly, getConnectionMode } from "./connection.js";
@@ -9,5 +10,6 @@ export {
   ensureSyncTables,
   ensureTradeDataTable,
   ensureReportingDataTable,
-  ensureMarketDataTables,
+  tableExists,
 } from "./schemas.js";
+export { ensureMarketTables } from "./market-schemas.js";
